+++ fs/reiserfs/objectid.c
@@ -21,7 +21,7 @@ static void check_objectid_map (struct super_block * s, __u32 * map)
 {
     if (le32_to_cpu (map[0]) != 1)
 	reiserfs_panic (s, "vs-15010: check_objectid_map: map corrupted: %lx",
-			le32_to_cpu (map[0]));
+			( long unsigned int ) le32_to_cpu (map[0]));
 
     // FIXME: add something else here
 }
@@ -161,7 +161,7 @@ void reiserfs_release_objectid (struct reiserfs_transaction_handle *th,
     }
 
     reiserfs_warning ("vs-15010: reiserfs_release_objectid: tried to free free object id (%lu)", 
-		      objectid_to_release);
+		      ( long unsigned ) objectid_to_release);
 }
